requirements.txt
tkcalendar
pandas
getmac
pyaudio
pyserial
pysound
pytest
cryptography
customtkinter
CTkMessagebox
CTkMenuBar
playsound
packaging
requests
pyyaml
